Localize the working container name to 'create'
The working container and its name is a temporary and internal detail
of the 'create' command. It's not something that someone would want to
manually tweak.

Since 'local' is not mandated by POSIX, use a subshell for the
function instead.

This is a step towards making the commands more parametrized so that
one can use different toolbox containers.