Use the host's PID namespace for the toolbox container
It's common practice to track down a process while developing. This
could be a process that's repeatedly crashing, or something that's
misbehaving by using up too many resources. Being able to seamlessly
look at what's happening on the host makes for a better developer
experience.

The toolbox doesn't intend to provide a segregated security domains,
so this is fine.