Use sed(1) instead of tr(1) for consistency
Shell scripts are archaic enough by themselves. For the sake of
readability it's better to stick to a smaller subset of features
that's already being widely used, instead of introducing slightly
different ways of doing the same thing.

Unless there's a pressing user-visible reason to introduce tr(1), it's
better to keep using sed(1).