Avoid a Bash-ism (ie. echo flags) and stick to POSIX (ie. printf)
Detected by https://www.shellcheck.net/.

Note that for coloured output the escape sequences need to be
interpreted as control characters and hence the variables containing
them need to be expanded inside the format string. This is an exception
to the general rule of thumb regarding printf [1] and hence SC2059
needs to be silenced. Otherwise https://www.shellcheck.net/ would
complain:
  Line 740:
  printf "${LBC}Images created by toolbox${NC}\n"
         ^-- SC2059: Don't use variables in the printf format string.
           Use printf "..%s.." "$foo".

Some changes by Debarshi Ray.

[1] https://github.com/koalaman/shellcheck/wiki/SC2059

https://github.com/debarshiray/toolbox/issues/53