Support base toolbox image names that include a domain
A subsequent commit will leverage this to let the user override the
base toolbox image with an image reference that might include a domain.
This can be helpful if the user wants to use an image from a specific
registry as opposed to the current approach of first looking for
localhost/<image> and then registry.fedoraproject.org/f$release/$image.

The format of a image reference is defined in:
https://github.com/moby/moby/blob/master/image/spec/v1.1.md
https://github.com/docker/distribution/blob/master/reference/reference.go

The code in image_reference_has_domain is based on the
splitDockerDomain function implemented in:
https://github.com/docker/distribution/blob/master/reference/normalize.go

https://github.com/debarshiray/toolbox/pull/73