Try to avoid downloading a base image that's referred to by its ID
When any image is referred to by its non-human-readable ID, then it
must already be cached locally. It cannot be pulled from a registry
based on the hash, and there's no way to map it to the human-readable
name either.

A subsequent commit will leverage this to let the user override the
base toolbox image with an image reference that's a non-human-readable
ID.

The format of a image reference is defined in:
https://github.com/moby/moby/blob/master/image/spec/v1.1.md
https://github.com/docker/distribution/blob/master/reference/reference.go

https://github.com/debarshiray/toolbox/pull/73