Provide meaningful hint to overcome breakage caused by Podman 1.2.0
Podman commit 4a02713c57d874c4 [1] broke 'podman exec' for existing
containers.

Note that support for ConmonPidFile in 'podman inspect' was added in
the same version of Podman as the one that has the breakage caused by
commit 4a02713c57d874c4 [1]. Therefore, there's no need to bump the
minimum required version of Podman. Interim Git snapshots between the
two will only cause the extra hint in the error message to be hidden.

[1] https://github.com/containers/libpod/commit/4a02713c57d874c4

https://github.com/containers/libpod/issues/2673