Add c.gh.debarshiray.toolbox to containers and images from 'create'
Commit 0ab6eb7401fb3964 introduced the com.github.debarshiray.toolbox
label as a better way to identify toolbox containers and images. It
might be necessary to support creating toolbox containers from base
toolbox images that are not appropriately labeled. eg., this would
allow users to use newer versions of the toolbox script with older base
images to debug problems that are only seen in those versions of the
images.

While no change has been made, yet, to either enforce the presence or
override the absence of the com.github.debarshiray.toolbox label it is
prudent [1] to ensure that artifacts produced by 'toolbox create' are
adequately labeled regardless of what labels the input base image had.

[1] https://en.wikipedia.org/wiki/Robustness_principle

https://github.com/debarshiray/toolbox/pull/91