Include containers tagged with c.gh.debarshiray.toolbox in 'list'
First we need to get the containers with an old label and then with the
new label. We have to change the format of the query to only include the
name of the containers and not the other informations like 'created' as
they could make a problem when removing the duplicates with uniq. When
we have the final list of containers we pass them to
containers_get_details() where we obtain the final details for them.

https://github.com/debarshiray/toolbox/pull/101