Use a lighter entry point than /bin/sh
Now that toolbox containers no longer use a separate PID namespace [1],
the entry point specified in 'podman create ...' doesn't act as PID 1
inside the toolbox container. It's just a process that's spawned by
'podman start' to denote the state of the container. This opens the
possibility of using something even more lightweight, such as
'sleep +Inf'.

sleep(1) takes 64 kB compared to the 432 kB taken by /bin/sh.

This wouldn't have been possible with a separate PID namespace. In that
case, the entry point would also become PID 1, and since the only
signals that can be sent to a PID 1 are those for which it has
explicitly installed handlers, this would cause various problems.

[1] Commit 67522f0ad7637768

https://github.com/debarshiray/toolbox/pull/108