Always use a tag for the user-specific customized image
A subsequent commit will generate the name of the toolbox container
from that of the user-specific customize image, instead of just using
the same exact string. This is necessary because, unlike OCI images,
names of OCI containers can't have a colon in them [1]. Knowing that
the user-specific customized image always has a tag will make things
simpler.

Buildah and Podman are going to use 'latest' as the placeholder when no
tag is specified, and this what is explicitly mentioned when the base
toolbox image is referred to by its non-human-readable ID.

Fallout from 56c3cfc27c966dede6454eb13a93e6024d3c7f1c

[1] Podman commit 449b8ab7b14fcc0d
    https://github.com/containers/libpod/pull/2793

https://github.com/debarshiray/toolbox/issues/106