Make entrypoint=sleep, rather init container on exec
Change the entrypoint to be just a thin wrapper for `sleep infinity`;
we initialize the container on-demand the first time `exec` is
run, using a lockfile to serialize initialization across multiple
exec invocations.

This closes a race between start + exec and further execs, and
also paves the way to use fd passing for state.