Use both su and setpriv
`su` runs through PAM and sets up the environment (e.g. `HOME`).
But we need to drop the default capabilities we inherited, otherwise
our unpriv user will still have things like `CAP_SYS_ADMIN` (in the userns)
which is wrong - we wnat the user to feel unprivileged in the ns.