Fix typo
It was working because 'toolbox_container' is a global variable.
However, given that the name of the toolbox container is already being
passed as an argument to the function, it's better not to use the
global variable.

Fallout from c492907c12bed9b3c3556ea44a54dd321d5c6cfd

https://github.com/debarshiray/toolbox/pull/201