Ensure correct toolbox.sh is used by container
This fix an issue where /etc/profile.d/toolbox.sh was mounted inside the
container, which may happen with containers created with some previous
versions of toolbox, and then when entering the container toolbox would
not copy the latest version of the file to the container and the mounted
version would be used instead.

Without this change, some issues fixed (e.g. line wrapping on long
input [1]) in the current /etc/profile.d/toolbox.sh could not be applied
to these containers depending on the version of it that was mounted in
the container.

[1] Commit bc1a816ea3b80fc7c3dc5d523838256e94c953d2
    https://github.com/debarshiray/toolbox/issues/190

Signed-off-by: Andre Moreira Magalhaes <andre@endlessm.com>