Fix duplicates showing in list output
This removes an additional step when listing info about containers and
images. Until now we were filtering output of podman ps to get names of
containers that came from toolbox and then we got info about them. But
the lack of regex support in podman caused duplicates showing when there
was a "." present in the name of container. So this removes this step
and simply gets all the data in the first step. This is possible thanks
to the format flag in 'podman ps' that allows multiple columns to be
requested.

This also fixes possible duplicates of images due to a trailing space
behind a variable.