Make /srv, /mnt, /var/home consistently mount points
Anywhere the `/host/` leaks into tooling it can create a problem
because that path diverges inside and outside the container.

I keep running into problems with using tools like VS Code and
in particular golang and `$GOPATH`, both of which tend to want
absolute paths.

This way, my data in `/srv/walters` is at the "same place" (path)
both inside and outside the toolbox.