Split runtime and dynamic state
Currently we run with `--rm` which means "delete image when container removed".
However, when the host system reboots, today the default is for
containers to be *stopped* but persist.  This has interesting
semantics; the main tricky problem is that the filesystem persists,
but things in `tmpfs` don't.

Split up our initialization to distinguish between changes to persistent
storage versus runtime state such as mounts.

This makes `podman stop coreos-toolbox; coretoolbox run ...` work.