Optionally bind $HOME/.cache/toolbox/f[release] to /var/cache/dnf
Downloading dnf repository metadata for the first time in a new
container is expensive.  Fedora and Fedora Updates are 70mb and 19mb
when compressed, respectively, and the former does not change after
the distribution is released.

This change allows containers to share their dnf metadata cache,
potentially saving substantial time (and data) when creating
new containers, while allowing one to opt-out of the sharing
via a new "--private-dnf-cache" flag.