Use tmpfs for /tmp, and forward /var/tmp to host
Doing a complex dance with the tempdirs didn't quite work out;
in particular a problem was that the symlink to a temporary tempdir
was retained across host restarts, but the tempdir was lost.

I think we do need a private `/tmp`, so let's do that with
a `tmpfs`, same as the host.

However the use case for a private `/var/tmp` seems less strong;
let's just symlink that to the host.